refactor: 💡 add desktop pagination component
