+++ src/Molecules/Pagination/Pagination.types.ts
@@ -5,7 +5,18 @@ export type Props = {
   onPageChange: (newPage: number) => void;
 };
 
-export type PageNumberItemProps = { active?: boolean; onClick?: (e: React.SyntheticEvent) => void };
+export type DesktopPaginationProps = {
+  currentPage: number;
+  numberOfPages: number;
+  onClickPageNumber: PageNumberItemProps['onClick'];
+  onClickPrevious: ChevronButtonProps['onClick'];
+  onClickNext: ChevronButtonProps['onClick'];
+};
+
+export type PageNumberItemProps = {
+  active?: boolean;
+  onClick?: (e: React.SyntheticEvent) => void;
+};
 
 export type ChevronButtonProps = {
   direction?: 'left' | 'right';
